More test to push file from my computer to gihub
After link the projets: in my computer and Gihub (they need to have the same name)

After everything chosed, do.

Seacher in my CP the <project>.Rproj "file"
click on it
Is going to open Rstudio that is alredy linked to the projcet in Gihub

After in Rstudio: Envirronment window select the file that is going to be push to Gihub: commit and Push. Done.